Set default libcore NetworkSecuriyPolicy for DownloadManager
The RealSystemFacade that applies the network security config for the
download manager shared process does not set the default NSPolicy
object. Therefore, some security settings like CT enabled/disabled are
not applied correctly when a download is started on behalf of an app.

Also, the NetworkSecurityPolicy object is a singleton and depends on the
app NetworkSecurityConfig. If the DownloadManager starts multiple
downloads in parallel on behalf of different apps, the NSPolicy object
might become inconsistent. As a fix, this CL imposes that downloads are
performed sequentially, to avoid race conditions in the setting of the
default NSPolicy.

Bug: 407952621
Test: atest CtsNetSecConfigDownloadManagerTestCases
Flag: com.android.org.conscrypt.net.flags.certificate_transparency_default_enabled
Change-Id: I06017bf38488aeafbd2906b9086db7ac64d326fa